refactor print functions to output_service
